[Editor] Remame checker to solver, fix bug in importer [TI] Fix scroll bug. Increase maximal number of level.

+++ IsekaiTI83-84/src/loader.h
@@ -13,7 +13,7 @@ extern uint8_t num_packs;
 void load_packs_info();
 void free_packs_info();
 
-void load_level_data(pack_info_t pack, uint8_t level);
+void load_level_data(pack_info_t pack, uint16_t level);
 
 void save();
 void load_save();
